+++ lib/ui/WhatsAroundMe/wam_list_model.dart
