+++ lib/ui/WhatsAroundMe/wam_nearby_search_model.dart
